\name{boxplotall}
\alias{boxplotall}
\title{Create boxplots of all quantitative variables using all factor variables}
\usage{
boxplotall(data, no.export = TRUE, ...)
}
\arguments{
  \item{data}{the data frame to use}

  \item{no.export}{if this is set to false, then graphs
  will be exported to png}
}
\value{
null
}
\description{
Create boxplots of all quantitative variables using all
factor variables
}

