\name{chrisplotall-package}
\alias{chrisplotall-package}
\alias{chrisplotall}
\docType{package}
\title{
chrisplotall
}
\description{
ery simple functions to plot all possible scatterplots, boxplots, and histograms in a data frame. Useful for exploratory analysis.
}
\details{
\tabular{ll}{
Package: \tab chrisplotall\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2013-12-12\cr
License: \tab WTFPL-2.0\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Christopher Beckham

Maintainer: <chrispy645@gmail.com>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
