\name{scatterall}
\alias{scatterall}
\title{Create scatterplots out of all possible quantitative variables in the data.}
\usage{
scatterall(data, no.export = TRUE, ...)
}
\arguments{
  \item{data}{the data frame to use}

  \item{no.export}{if this is set to false, then graphs
  will be exported to png}
}
\value{
null
}
\description{
Create scatterplots out of all possible quantitative
variables in the data.
}

